chore: refactor crud services
