+++ app/modules/cloud-code/index.js
@@ -1,0 +1,26 @@
+import {
+  customMultiParamQueryService,
+  customQueryService,
+  getObjectsByGeolocation,
+  postObjectsToClass,
+  postObjectsToClassWithRelation,
+  removeQueryService,
+  residentIDQuery,
+  updateObject,
+  countObject
+} from './crud';
+import { retrieveCustomData, retrieveUniqueListOfOrganizations } from './custom-data-retrieval';
+
+export {
+  customMultiParamQueryService,
+  customQueryService,
+  getObjectsByGeolocation,
+  postObjectsToClass,
+  postObjectsToClassWithRelation,
+  removeQueryService,
+  residentIDQuery,
+  retrieveCustomData,
+  retrieveUniqueListOfOrganizations,
+  updateObject,
+  countObject
+};
