--- app/modules/parse/index.js
@@ -1,3 +1,0 @@
-import { retrieveCustomData, retrieveUniqueListOfOrganizations } from './data-retrieval';
-
-export { retrieveCustomData, retrieveUniqueListOfOrganizations };
